allowed editing of both name and label for terms, refactored interfaces
