+++ src/components/category/CategoryForm.tsx
@@ -1,23 +1,6 @@
 import React from 'react';
 import UniversalForm from '@/components/framework/UniversalForm';
-
-type FormChangeEvent =
-  | React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
-  | { target: { name: string; value: string } };
-
-interface CategoryFormProps {
-  form: {
-    name: string;
-    code: string;
-    description: string;
-  };
-  onChange: (e: FormChangeEvent) => void;
-  onSubmit: (e: React.FormEvent) => void;
-  error?: string | null;
-  success?: string | null;
-  isEditMode?: boolean;
-  onCancel?: () => void;
-}
+import type { CategoryFormProps } from '@/interfaces/CategoryInterface';
 
 const CategoryForm: React.FC<CategoryFormProps> = ({
   form,
