+++ src/interfaces/AssociationInterface.ts
@@ -1,3 +1,5 @@
+import type { Category } from '@/interfaces/CategoryInterface';
+
 export interface Association {
   name: string;
   identifier: string;
@@ -6,4 +8,25 @@ export interface Association {
   status: string;
   category: string;
   index?: number;
+}
+
+export interface AssociationCategoriesProps {
+  categories: Category[];
+  termName: string;
+  categoryName: string;
+  onBadgeClick: (
+    categories: Category[],
+    termName: string,
+    categoryName: string
+  ) => void;
+}
+
+export interface AssociationDetailsModalProps {
+  open: boolean;
+  onClose: () => void;
+  categories: Category[];
+  termName: string;
+  categoryName: string;
+  expandedCategory: string | null;
+  onCategoryClick: (id: string) => void;
 }
